cleanup + reorganize folder
